requirements.txt
Django==3.2.23
djangorestframework==3.14.0
pandas==2.0.1
geopy==2.3.0
requests==2.31.0